chore: mv from_env to ClientConfig type
